cleanup initialization and freeing to alleviate some crashing and other breakage.
ob3ds should either belong to the node for their entire lifetime, or belong to the parent context. currently only the node and the videoplane belong to the parent, all other objects belong to the node. only the node and vplane need to allow modifying the drawto.

also, explicitly free all ob3ds, seems to help issues with multiple instances of meshwarp abstraction loaded.